requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
supabase==2.1.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0 